chore: add missing integration test prerequisite (#361)
* chore: run build step before `npm run test`

The integration tests require the source to have been built.

The e2e tests do as well, but are already running this in a different way.

There isn't much of a clean separation of the test config by test type, so I decided to keep it
simple by just adding it as a step on a full `npm run test` run. Technically you could still pull
down this repo and start running `vitest` and run into this, but that seems rare enough and I'll
document this prerequisite in my next commit.

* docs(readme): mention prereq for integration tests

See previous commit. Ideally we'd fully automate this but it may not be worth the trouble at the
moment.