chore(e2e): handle and surface retries + more fixes (#567)
* chore(e2e): fix parsing of JUnit XML file numbers

These are all returned as strings, so it's easy to introduce bugs downstream if we don't immediately
convert them explicitly to numbers.

This fixes some cases where we render `NaN` on the report page.

* chore(e2e): fix rendering of skipped tests in main section

If all the tests in a suite were individually skipped, we ended up with an empty section and a `Nan%`.

* ci(e2e): fix default for workflow_dispatch versions input

* chore(e2e): handle retries in next.js e2e results

Remove all suites known to be fully passing from the `e2e-skip-retry.json`, add logic to handle
retries in the test results, thread this through to the final results, and show retry info on the
report page.

Every suite was added to this list in order to tame the test run time, but now that we're around a
98% pass rate, the impact of enabling retries should be minimal.

* refactor(e2e): simplify and clarify junit2json skip handling

* chore(e2e): don't duplicate tests skipped via config

These were ending up inserted because they're skipped and inserted a second time if they also ran
and failed. The same could technically happen if it passed and was skipped.