chore(e2e): still more e2e report fixes (#578)
* chore(e2e): fix e2e result deduping for next@13.5.1

Example broken run: https://github.com/netlify/next-runtime-minimal/actions/runs/9410953045.

* chore(e2e): fix summary counts when deduping occurs

We were updating summary tallies as side effects of building up the report, but when I introduced
deduping of test suite retries, this led to a discrepancy between the results in the report and the
summary counts.

I could have fixed this by subtracting the right counts from the running tallies when deduping, but
it seemed cleaner to just refactor this to summary counts computed purely from the final results.

As a bonus, this makes it a lot more obvious that the skipped count makes no sense. I translated the
logic as is, but I'll try to fix this in the next commit.

* chore(e2e): more explicitly output skip counts

It was really misleading and error prone.

This also removes the total, since it wasn't necessary and it was also misleading. We were
calculating it as `passed + failed + individually skipped tests` and then only using it in one
place, which subtracted the individually skipped tests back out.

This also updates the local dev fixtures.