fix(edge-runtime): match path with URI-encoded chars (#2873)
Following the behaviour in https://github.com/vercel/next.js/pull/78325, if a path
matches a configured middleware matcher when URI-decoded, consider it a match.

This doesn't seem ideal, but it aligns our behaviour with that of Next.js, and ensures
that static asset matching behaviour aligns with edge middleware matching behaviour.

Co-authored-by: Mateusz Bocian <mrstork@users.noreply.github.com>